Improve CI testing. (#13)
The envoy-filter-example CI did not actually verify that a runnable
binary was built since it only ran a test. A change in Envoy resulted in
the binary build from the envoy-filter-example becoming broken.

The submodule hash is updated as part of this commit.  The WORKSPACE
file is modified to include references to the api dependencies added
since the last submodule hash update.

After this is merged a commit to Envoy will update the filter-example
hash for CI to include this commit, and modify do_ci.sh to run this
new test.